+++ frontend/plugin/src/lib/models/expression.ts
@@ -1,0 +1,15 @@
+export interface ExpressionFunctionInfo {
+  name: string;
+  description: string;
+  overloads: OverloadInfo[];
+}
+
+export interface OverloadInfo {
+  arguments: ArgumentInfo[];
+  returnType: string;
+}
+
+export interface ArgumentInfo {
+  name: string;
+  type: string;
+}