+++ frontend/plugin/src/lib/models/index.ts
@@ -1,3 +1,4 @@
 export * from './async-resource';
 export * from './config';
+export * from './expression';
 export * from './template';